pyproject.toml
[tool.poetry]
name = "webradius"
version = "0.1.0"
description = ""
authors = ["Zarembo Pavel"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.5"
configparser = "^5.3.0"
aart = "^0.0.1"

[tool.poetry.group.dev.dependencies]
Django = "^4.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
